Initial commit
This version was already published as "r2: 2016-05-20" on
http://liederbuch.frederikheld.de/

There might have been a local git repository on my old computer
but that is not accessible right now and might not be very
interesting.

This commit includes the following three third-party libraries:
- FontAwesome
- jQuery
- snap.js
They all have licenses that are compatible to the GNU GPLv3 used
in this project.

The sheets are not published with this repository since the
copyright of those songs might be in the hands of others.